+++ eyes.sdk.core/src/main/java/com/applitools/eyes/TaskListener.java
@@ -1,6 +1,0 @@
-package com.applitools.eyes;
-
-public interface TaskListener<T> {
-        void onComplete(T taskResponse);
-        void onFail();
-}